Create README.md
This repository contains full posteriors for both instantaneous and lifetime integrated X-ray fluxes as a function of stellar mass and age (calculated in McDonald, Kreidberg, & Lopez 2019). The data are stored as 3-dimensional numpy arrays. 

The "inst_" files are instantaneous X-ray flux as a function of age and stellar mass. The "posterior_" files are lifetime-integrated X-ray flux as a function of age and stellar mass. The number in the filename, e.g. "175" is the metallicity of the specific table (equal to Z = 0.0175) and, there are 5 different metallicities. 

FIXME: describe what the three dimensions are.